Fixing some incorrect member declarations.
git-svn-id: svn://svn.renegadeprojects.com/yrpp/trunk@333 bbc44d77-a14d-0410-bbfb-905795908d03

+++ RulesClass.h
@@ -356,7 +356,7 @@ public:
 	int				 IonCannonDamage;
 	int				 RailgunDamageRadius;
 	BuildingTypeClass*		 PrismType;
-	float				 PrismSupportModifier;
+	int		PrismSupportModifier; // this is int(100 * ini value)
 	int				 PrismSupportMax;
 	int				 PrismSupportDelay;
 	int				 PrismSupportDuration;
